# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StudyRok Backend API
  version: 1.0.0
  description: REST API for authentication, study boards, materials, generated notes, flashcards, quizzes, and background generation progress.
  contact:
    name: StudyRok Support
servers:
  - url: http://localhost:3000/api
    description: Local API server
  - url: http://localhost:3000
    description: Local server root
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: string
                        description: New access token.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ApiResponse:
      type: object
      required: [success, statusCode, message, timestamp]
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        message:
          type: string
        data:
          description: Response payload, when present.
        meta:
          description: Optional response metadata.
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [success, statusCode, message, timestamp]
      properties:
        success:
          type: boolean
          example: false
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: object
          properties:
            code:
              type: string
            details:
              description: Optional structured error details.
        timestamp:
          type: string
          format: date-time
...